including a log entry & fix

+++ controllersvc/src/main/java/com/emc/storageos/volumecontroller/impl/ecs/ECSObjectStorageDevice.java
@@ -34,6 +34,7 @@ import com.emc.storageos.model.object.BucketACL;
 import com.emc.storageos.model.object.BucketACLUpdateParams;
 import com.emc.storageos.svcs.errorhandling.resources.InternalException;
 import com.emc.storageos.volumecontroller.ControllerException;
+import com.emc.storageos.volumecontroller.ObjectControllerConstants;
 import com.emc.storageos.volumecontroller.ObjectDeviceInputOutput;
 import com.emc.storageos.volumecontroller.ObjectStorageDevice;
 import com.emc.storageos.volumecontroller.impl.BiosCommandResult;
@@ -142,11 +143,15 @@ public class ECSObjectStorageDevice implements ObjectStorageDevice {
     }
 
     @Override
-    public BiosCommandResult doDeleteBucket(StorageSystem storageObj, Bucket bucket, final String taskId) {
+    public BiosCommandResult doDeleteBucket(StorageSystem storageObj, Bucket bucket, String deleteType, final String taskId) {
         BiosCommandResult result;
         try {
             ECSApi objectAPI = getAPI(storageObj);
-            objectAPI.deleteBucket(bucket.getName(), bucket.getNamespace());
+            if (ObjectControllerConstants.DeleteTypeEnum.VIPR_ONLY.toString().equalsIgnoreCase(deleteType.toString())) {
+                _log.info("Inventory only bucket delete {}", bucket.getName());
+            } else {
+                objectAPI.deleteBucket(bucket.getName(), bucket.getNamespace());
+            }
             // Deleting the ACL for bucket if any
             List<ObjectBucketACL> aclToDelete = queryDbBucketACL(bucket);
             if (aclToDelete != null && !aclToDelete.isEmpty()) {
